tg-bot/requirements.txt
python-telegram-bot==20.0
psycopg2==2.9.10
SQLAlchemy==2.0.40
httpx~=0.23.3
alembic~=1.16.1
redis==4.5.0